+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+### 1.7.8
+
+- Add module settings for remove popout from the requestor module asked from https://github.com/OhhLoz/Harvester/issues/37
+
 ### 1.7.6-7
 
-- Fix wrong "formatDragon" import from https://github.com/OhhLoz/Harvester/issues/37#issuecomment-1981881466
+- Fix wrong "formatDragon" import from https://github.com/OhhLoz/Harvester/issues/37
 
 ### 1.7.4-5